requirements.txt
flask==3.1.0
openai-agents>=0.0.7
python-dotenv>=1.0.0
